requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
exceptiongroup==1.2.2
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
numpy==2.0.2
packaging==24.2
preshed==3.0.9
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-docx==1.1.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.41.3
thinc==8.3.2
tqdm==4.67.0
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
